feat(star wars): add test fixtures for intergration tests.
The fixtures for testing `counterfeit` promises have been added.
These fixtures consist of a module called `starWars`, that includes the
following classes:

* DeflectorShield
* DeathStar

The `DeflectorShield` service exposes one public method called `reboot`.
This asynchronous operation returns a promise.

The `DeathStar` factory exposes a public method called
`rebootDeflectorShield`. This operation delegates to
`DeflectorShield#reboot` and interfaces with the promise that is
returned.