feat(counterfeit): add api for creating promises and stubs
A factory called `counterfeit` has been added. This factory facilitates
the testing of methods that use asynchronous promises, by providing an
api to generate fake promises and stub functions that return a promise.

An integration test has been added for the `DeathStar` test fixture.
Within this test the `DeflectorShield` class is decorated so that the
`reboot` method is a `CounterfeitStub`. This stub is configured to
return a `CounterfeitPromise`, which can be conveniently resolved within
the test.

Behind the scenes `CounterfeitPromise` are built using the AngularJs
`$q` service.