feat(CounterfeitPromise): add raw property
A new public property `raw` has been added to `CounterfeitPromise`. This
property references the promise object associated with the internal
deferred object.

Developers may find it useful to use `raw` if they have a requirement
to override an object property that is a promise in a testing
environment. Contrast this to the requirement to stub a function that
returns a promise, where you would use `counterfeit.stub`.